formsflow-admin-api/requirements/prod.txt
gunicorn
Flask
Flask-Caching
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy
flask-restx
flask-jwt-oidc
python-dotenv
psycopg2-binary
jsonschema
requests
attrs
dpath
Werkzeug
cryptography
sqlalchemy_utils
sqlalchemy
Inflector
PyJWT
